move to org.eclipse.microprofile package
for shutting up the discussions and getting back to productive work again
