+++ tck/src/main/java/org/eclipse/microprofile/config/tck/ConfigProviderTest.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  *
  */
-package io.microprofile.config.tck;
+package org.eclipse.microprofile.config.tck;
 
 import java.util.Map;
 import java.util.Properties;
 
-import io.microprofile.config.Config;
-import io.microprofile.config.ConfigProvider;
+import org.eclipse.microprofile.config.Config;
+import org.eclipse.microprofile.config.ConfigProvider;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
